This is a set of changes provided by David Mark.  His updates were based on old versions of these files, so I merged the deltas by hand.  My apologies if I accidentally trounced an old patch.
David did not provide a change log, but most of these changes seem to be improvements to the tk interface.

One item that David pointed out that is important is to use the enclosed copy of event_sounds.txt.  The default mh.ini parameter points to $Pgm_Root/data/event_sounds.txt but if you override this parameter, be sure to use this version of the file.