Updates to iButton (1-wire) support from John Murphy to support DS2409 based 1-wire hubs.  The following are comments from John:
I wrote this a while ago, not sure it works for anyone else, but it searches sub branches on 1-wire networks for 1-wire couplers and searches those sub branches.  It does the discovery on the first attempt to access a specific serial number, then keeps a record, so polling all devices is done once the discovery is completed.  I use mine with the AAS hub that has 3 DS2409 couplers in it.

These should drop in and work just like the regular 1-wire files/items in MH, the DS2409 couplers are taken care of under the hood so to speak, so no configuration changes to MH are needed.