Fixed logic in &state.  If $@ was set upon entry to &state, then a Weather_Item eval error would be printed if an undefined Weather item was being evaluated and $valid was set to 1.
In other words, a Weather_Item eval error would be printed even if it was a previous eval error that caused $@ to be populated.

In the reworked logic, the print statement is only reached if an eval statement is actually executed and an error is generated by that eval statement.