r1822 by Gregg attempted to clarify, through the name of voice commands, whether commands apply to the PLM or to all devices. For example, we'd have "this: show link table to log", or "all: scan link tables" in the PLM menu in the "Insteon" category in the web interface. The problem with adding "this:" and "all:" to the voice command names is twofold:
1. The Tk interface fails to load because the ':' character has special
meaning there, so adding a ':' to a command name confuses some part of
the Tk interface, causing it to fail to load.

2. Voice commands become awkward since now we'd have commands like:

"plm this: show link table to log"

"plm all: scan link tables"

For now I am reverting the changes that add "this:" and "all:" to voice
commands so the Tk interface loads again. We still can use a character
different from ':', e.g. use "this-" and "all-", and that'd solve (1)
above, but would still leave (2).

Will leave up to Gregg to decide when he gets a chance whether he wants
to keep "this" and "all" in the voice command names.