Reverted a commit that broke the xPL listening socket on some platforms
In a previous commit, a piece of code was introduced that should
take care of selecting a free port automatically by passing port
0 to the open_port function.

Some people reported problems with that code change, especially on ubuntu
so I reverted that change (and only that one, the other changes were not
touched).

I verified that the code change did not work on Ubuntu 12.04 and that this commit
does work fine on the same platform.