Reverted the change to xPL_Items.pm
The device_monitor function was changed in a previous commit to match any characters instead of only non-whitespace characters. This causes some existing xPL device stop funcitoning with MH, as state in #10. What was the reasoning behind this change?

I guess it doesn't make sense to make MH restrictive in what messages is can process. IMHO whitespace should not cause a message to fail, as xPL is supposed to be a human-readable format.