Fixed TTS on OS X
The previous version of speak_mac depended on a depricated module
called Mac::Speech. This module no longer works on 64-bit installs.

I've replaced the TTS code with a simple call to the buil-tin command
'say'.
It uses the default system voice and volume.

Disadvantage: it only works to produce TTS on the local machine. Rerouting
the sound to other devices needs further discussion.