Allow for Hop Count of 0 and Set Initial Hop Count to 0
The lowest hop count permitted by Insteon is 0 not 1.  A hop count of 0 means that the PLM will send the message, but no other devices are to repeat the message.  Generally devices near the PLM will not require any hopping.

-Fixed code to allow for setting a device's hop count to 0.
-Added message flags for hop counts of 0
-Set the initial hop count for new devices to 0

It is important to be judicious with increasing hop counts.  As hop counts increase, more messages are sent around the network which can lead to unintended interfence.  In addition, increasing the hop count increases response times.  This is because the responding device will not respond until all possible repeats of the message have occurred.  Even if a device hears the message on the first broadcast, it will wait for all hops to occur before responding to avoid collisions.