Prevent Adding a Link if ALDB is Out of Sync
Changes made to ALDB_i1 Add_link so it will prevent any adding of a link from any voice command.  The code looks like crap, references and callbacks stacked on top of references and callbacks.  What a mess.

Should be able to do the same thing for update and delete using the same method.