Only Mark Duplicate in ALDB if address is not the same
When scan all links, MH ALDB database is cleared.

But for scan_one which only scans a single address, we don't dump the ALDB database.  So the fact that an aldbkey already exists does not make it a duplicate, we could be scanning the same address again.  It is only a duplicate if the addresses are different.