Bump ALDB_Delta if Set to 00
The ALDB_Delta gets reset to 00 when devices lose power.  The ALDB_Delta also gets reset to 00 when the devices are factory reset.

Since we can't distinguish between these two states, we never allow the ALDB_delta to stay at 00.

If a devices loses power, the ALDB_delta will change and a full scan will be conducted.
At the conclusion of the scan we attempt to query and store the aldb_delta.
If the aldb_delta is 00 we find the first empty address on the device.
We then peek the aldb_flag(the first byte) of this address
We then poke the same value back into the device.
We use an empty address to decrease the potential harm from writing an incorrect byte if something goes wrong
We then scan this address, being empty, the scan should end after the first byte
We then query the ALDB_delta which should now be 01 since we "wrote" a single byte.  The value increases even though the data didn't change.

I reverted the previous commit as I was headed down the wrong path.

Solves krkeegan/misterhouse#3