Ignore Peek and Set_msb Replied Where Cmd1 is Invalid
Fixes hollie/misterhouse#56

Responses to Peek requests should contain command code 2b as Cmd1, if we receive a message in which Cmd1 is not 2b, then it could either be corrupted or an out of sequence message.  To prevent corrupting MH link hash, ignore the message.  This will cause a resend to occur.

Similarly Set_msb responses should contain command code 28 as Cmd1.  Do the same if Cmd1 is incorrect.

This logic probably would be more appropriate in the _on_peek portion of AllLinkDatabase.pm, however, _on_standard_receive blindly clears the active message without concern for whether the messages are valid.  As such, the logic had to be placed at the lower BaseInsteon.pm level.