Set Default Hop Count to Highest Count in Hop History
Add hop_history function which adds a hop count to an array of a maximum length of 10.  Then takes the highest hop value from the history and sets it as the default_hop_count.

In addition, anytime a message needs to be resent, the default hop count is immediately bumped up by 1.  Previously if default_hops was at 2 it required 2 failures before bumping the hop count to 3.  The default hop count is bumped up by adding a hop count of that value to the hop history.

This results in the value of default_hop as being the highest hop count needed for the last 10 sent or received messages.