Reformulate Send_Timeout for All Link Commands
All Link Commands from the PLM can take a very long time to process.  The amount of time depends on the number of objects linked to this scene.  Plus if the PLM is forced to send any direct commands, this can take even longer.

For example, on my network, sending a command to 20 devices without any direct commands took 6 seconds.

We obviously can't set a send_timeout for this amount of time.  So instead, we set a normal initial timeout and then keep resetting the timeout when we receive a direct ACK or NACK because these are signs that the PLM is still working on the command.

Now, since the final message from the PLM should never fail to arrive, we never really need to worry about this.  But we need some timeout, otherwise we have the possibility of having MH lockup waiting for a message that never arrives.