Remove ACK Hash and Correct for Discrepancy in Documentation for ALL-Link Cleanup Failure Report
Misunderstood what would be warrant a NACK in the All Link Status Report.  A NACK is only issued if the PLM is interrupted in sending the direct commands. As such ACK HASH is not needed.

Instead an ALL-Link Cleanup Failure Report is issued if a device fails to respond.

However, the ALL-Link Cleanup Failure Report is incorrectly documented in the Modem guide.  The message is only 6 bytes long and byte 3 in the documentation doesn't exist.