Map aldb version based on restored engine_version
This update will correctly map the I2 aldb routines for devices that
were previously saved as I2 or I2CS just after restoring saved states.
The startup scan will also validate the aldb type if it is run but it
is no longer necessary.  When adding a new device the user must run
the 'get engine version' voice command once or use the startup poll.