Add ALDB_Delta to _on_peek
Bump_Delta is essentially the same as delete.  Except it wrote back the byte that was just peeked, possibily further decreasing corruption errors.  In order to decrease the complexity of the coding it was dropped.

Scan_One was also dropped.  It enabled the scanning of one memory address.  It was called after an add, delete, or update to confirm that the action was done correctly.  It also decreased MH issues.  In order to decrease the complexity of the code, it was dropped.