Create Thermostat Broadcast Dummy Object
The 2441th Insteon thermostat will broadcast changes to all devices registered as a responder to group EF.  Created a dummy object under thermostat that is registered as a subdevice on group EF of the thermostat.  Sync links properly handles adding the link to the device, but has an error adding it to the PLM.  PLM responder link is not necessary, but will solve in future commit.

Moved reclass of thermostat to i1 and i2 to Insteon.pm.  Will add subchildren to same sub in future.