Using VERSION to determine the MisterHouse program version.
I propose to use the VERSION file in the root of the MisterHouse codebase to determine what version of Misterhouse we're running.
This is opposed to the hardcoded version in bin/mh.

I've cleaned out some legacy code that tried to determine the svn version (we're no longer using that) and now bin/mh reads the VERSION file. For the master branch this file reads 'unstable' by default. When we make a new version now, we only need to update the version number in the VERSION file to have it pushed to all locations in mh that rely on or use the version number.