Modify list_groups_by_object to allow for non-recursive listing
list_groups_by_object identifies all of the groups that an object is a member of.  Currently, if an object is a member of a child group, the child and parent group will be returned in the list.

This change allows for only the groups which the object is a direct member of to be returned.