Insteon::Thermo_i2 Add subs to interpret status mode and temp
Insteon is so annoying, why can't it pick a single format for the same data?

Anyways, temp in status message is tempx2, instead of long temp form.

In addition, for reasons beyond comprehension, the mode in a status message conveys the same data, but uses a different flag than a extended get response.  So stupid.