Fixed #77
In #10 I reverted a change in the xPL message parsing because it apparently caused the xPL processing of messages to stop working. In #10 I had the impression that the change was required to allow MisterHouse to support unsupported messages to be processed.
However, I was wrong there. I misinterpreted the xPL spec. When a message is created, it contains key/value pairs. Only the name of a key/value pair cannot contain a space. The value can. This fix enables processing of messages with spaces in the value of key/value pairs to be processed again.