Fixed AppStatus for xPL items.
I want to be able to trigger an event when an xPL items goes down. I noticed I could use the AppStatus local variable for that. However, that variable is never set to
'dead' because of a typo in the local variable. Also, the location where the variable was set to dead was incorrect.

At the same time I've changes the 'eval' for the dead_action to dead_action->() because otherwise the dead action was never executed on my system.

Now I can do
$dummy_device->dead_action(\&handle_dummy_down);  #noloop

sub handle_dummy_down {
     print_log "Dummy is now down";
}

and it works! Yay!