Adding support for X10 basic commands over xPL
This pull request is based on an email that was sent on the list. The original developer is Roger Simon.

I only did basic testing on this code (as in: I added an item in my mht file and saw it showed up in the web interface and it didn't crash misterhouse).

I think this code is missing default on/off states for X10Basic items?