Fix Issue #132: Remove tracking scripts from user docs
docs/toc.html contained two tracking scripts which no one on the project recalls.  Scripts have been removed and can be added back in later if someone comes up with a patch that supports opt-in.  The way the docs directory works every local MH copy was running the scripts as was the toc.html on sourceforge.net.