Fix for issue #151 - Warnings in message decoder test
Corrected bug exposed by warning when running MessageDecoder_test.pl.  There are still two sets of messages that generate warnings.  Those messages are intentionally used because they are not defined in the message definition hashes.  It is reasonable to expect warnings given that those messages are not defined.  Additional check logic could be added to prevent the warnings and write out error messages but that seems like a little bit overkill for a debug tool.