Insteon_Thermo_i2: Remove Complex Code Creating Child Objects
The code is just too difficult to try and make work.  Either the child objects have to be stored in a hash, which results in awful looking veriable names that are not user friendly to the basic user.

The removed code made user friendly variable names, but these variables could not be used in other user code because the init code was not eval'd until runtime.