Insteon_IOLinc: Completely Revamp IOLinc Class
The sensor is a seperate "group" item on the device, it is capable of having its own links and will send alllink broadcasts when it changes state.  Therefore, the sensor should be defined as a seperate object.

The code as committed works, to get the status of either object simply call request_status() on the relevant object.

ToDo:
- Add voice commands for the sensor
- Add the ability to set the few bits that exist