Insteon: More Adjustments to is_duplicate_received
The message window is now calculated based on some basis in reality.  The message window is the amount of time during which no proper message should be received.

This depends on whether the message needs to be ACKed or not.

If a message needs to be ACKed, then no message should be received from the device until:
- All hops left expire
- Max hops + 1 elapses for an ACK to be received
- Max hops - less hops left for a new request to be sent

If no ACK is required then no message should be received from the device until:
-All hops left expire
-Max hops for a PLM message to be received by the device
(since MH is more sensitive to false positives in these types of messages, and since when duplicates occur here they normally occur very fast, we are more conservative for these types of messages)