Insteon: Redefine Active_Interface on Code Reload
Failing to redefine the active_interface leads to potentially two divergent versions of the same object.
Removed logic which required that active_interface be undefined before it could be defined.
This may cause issues if two PLMs are used in MH, although it is not clear to me that two PLMs were ever supported by MH or that such support is even needed.

Closed hollie/misterhouse#57