Insteon_Thermo: Add support for Groups 2-4, Add Status Child
Groups 1-4 of the i2 devices report on/off events for cooling, heating, high humidity, and low humidity.  Insteon objects can be linked to these events.  e.g. Turn on a fan when the humidity is high.

In addition, a new status child has been added.  It provides a simple text of the current status of the HVAC system.