Fix issues with INSTEON lights in web interface
INSTEON lights are treated differently in the web interface versus X10
lights, for example: they get different icons, and cannot be dimmed or
brightened. This patch does the following:

- Use the same light icons for INSTEON lights that are used for X10
lights.

- Allow dimming/brightening of INSTEON lights by clicking on the left
or right side of the icon in the web interface (only dimmable light, of
course).

- Fix a logic error that prevents disabling the button image cache. Ran
into this while creating new icons for my web interface.

- As an experimental "feature", and a small delay to try to give MH a
chance to receive the new state of an INSTEON device changed through the
web interface. This will allow the correct state to be displayed when
the web page renders again. Disabled by default since it's just a hack
(though it works good enough for me).

Comments welcome.