Insteon: Change "initiate linking as controller" Group to 00
I was having issues where some sort of corruption within the PLM was causing it to send messages to the 01 group, very annoying.

HouseLinc uses 00, plus the insteon documents seem to envision that 01 is a valid group that could be used for other purposes.

Group 01 links will not be deleted by delete_orphans to prevent causing issues for other users.