Better Formatting in Web Interface Print Log
Changed the web interface print log to a monospaced or fix width font and formatting.  This allows the nice formatting that relies on consistent spacing that has been visible in the text based logs, to be viewable online.  I realize that fixed width fonts are not always the prettiest, but I did the best I could.  If anyone has any suggestions for how to make it prettier, I am open to additional changes.

I had to work around one oddity.  The web code converts all \n characters to \n\r when formatting the html.  This is a blind conversion without any regard as to whether the \n character is within a <pre> tag.