Insteon_Stats: Track Hops in Duplicate Messages
They still provide some detail on how long messages take to arrive.

There may be one issue in that this may catch subsequent deliveries of the same message with a lower hops_left count.  Such as when the message is delivered on the 2nd and 3rd hop.  But at the same time, sometimes the messages that should take longer to arrive, as they have used more hops, somehow arrive first, so this may be a wash.  In the end, the best way to avoid this problem is to use proper hop counts.