Insteon: Pass Message Clearing Decision to on_read_write_aldb
on_read_write_aldb now returns a 1/0 corresponding to whether the current message should be cleared.

When a bad message arrived, on_read_write_aldb attempted to requeue the message that was currently pending.  However, _process_message did not clear the pending message until after this routine was run.  As a result, a new message was not queued because it was duplicative, but then the current message was cleared.  This resulted in stalling the message queue.

Fixes bug #258