Insteon: Convert Link_to_Interface to Handle i2cs or non-i2cs devices
The routine first checks to see if the device will respond to a get_engine_version request.
If any ACK is received the routine continues as normal.  In this case even if the device is an I2CS, the device must already have link entries as it responds to the PLM.
If a NAK is received, it is assumes that this device is an i2cs and the link_to_interface_i2cs routine is called.