Insteon: Reset ALDB Scan Time on Completed Successful Scan Only
Previously we reset when we started a scan

Now we reset when a scan has completed successfully and each time when the device tells us that the ALDB Delta has not changed.

We also reset if serial query_aldb_delta requests are received in less than 2 seconds.  This allows for sequentail aldb actions without slowing down to send a request to the device for its ALDB Delta when it is unlikely that it changed.