Insteon: Hijack is_Acknowledged for InterfaceController
InterfaceController is different in that when its state is set by an outside object, MH needs to update the linked devices.  By moving this to is_acknowledged, the states of the linked devices will only be updated if the scene command was successfull.

Is_responder also always needs to be set to true for interfacecontroller even though it is not a root device.