Insteon: Dont Use Data3 in PLM Linkkey
Data3 on controller links on the PLM should be set to the group.  The old code set this to 0 or the responding device's button.  It doesn't appear that this error had any major effect.  So rather than replace all of these links, simply ignore the data3 value in the linkkey so that old links will be accepted by sync and delete links.

The Data3 value of the responder links on the PLM should always be 00.

Therefore no concern that a different data3 value makes a link on the PLM unique.