Insteon: Catch, Report and Recover from Errors in Sync All Links
This enables the same feature that has been available in the scan_all_links function.  If an error occurs while attempting to sync_all_links, MH now recovers from the error, and continues to process the remainin objects in Sync_All.  At the conclusion of the Sync_All routine a message is displayed identifying the failed objects.

Partial Fix #73