Insteon: Convert DebugLevel Entries to Parent Device
ALDB objects have no real name, and as such, a user would not be able to add them to a debug line in the ini.  Moreover, if a user puts a device in the debug line, they likely expect to get all relevant messages, including ALDB messages for this device.

Removed debuglevel routine as it is unnecessary.