AD2: Don't Perform Wrap-Around Reset, Unless Highest Zone is Consistent
The alphanumeric fault messages are more complicated than i realized.  The list of faulted zones restarts at the lowest faulted zone everytime a fault occurs.  So for example, if zones 1-5 are faulted, and the panel has just finished sending the zone 3 fault message, if zone 7 is then faulted, the next message will be a zone 1 fault again.

If we blindly perform wrap-around resets, it is very easy to create a "fluttering" zone when multiple zones are faulted.

To cut down on the potential for fluttering, this change only performs a wrap-around reset if the highest faulted zone that is reported remains constant for 2 cycles.  While not completly ensuring that all zones have been report, it is unlikely that the panel will be interrupted at exactly the same point in its cycle twice.

The major downside of this, is that there may be a slight delay in reseting zones from fault -> ready.  This is particularly true of the higher unmapped zones.