AD2: Redesign Declaration of Wireless Items; Grandfather in Old Ini Declaration
The old design for the ini declaration was a little confusing.  Plus, carrying it over directly into the Configation Hash did not allow for mht style declarations of a wireless object.

The new design seems more logical to me and adds the framework to support an mht style declaration.

The old design for ini declarations is stil supported for backwards compatibility.