Insteon_PLM_Parser: Completely ReWritten _parse_data Routine
Changes the PLM parser to a First-In-First-Out style.  This should resolve the issue found in hollie/misterhouse#374 where shorter messages were being excerpted out of the middle of larger messages.

I believe the routine is significantly easier to read.

Removed the use of the split routine which created really odd results, and was not the most logical code to follow.

This also likely solves other latent issues that may not have been apparent.