Insteon_Irrigation: Remove Status Broadcast Code, Insteon PLM Does not Support
Every now and then I am reminded what an AWFUL protocol Insteon is.  This is one of them.

The EZFlora can be set to send out broadcast messages, however the PLM will not decode them.  As a result, MisterHouse never sees them, and they are useless.

According to the manual the PLM could be put into monitor mode allowing it to see these messages, however Insteon disabled that feature on the PLM.

It is beyond stupid, that the PLM will not display a broadcast message sent from a device in its own link table.  There is no security concern here.