Insteon: Allow Delete Orphans to Be Run on a Single Device.
Using Delete Orphans on a single device is not generally advisable.  Delete orphans works on the device level and not the link level, meaning it only cleans up a specific device, which may result in "half-links"

Now, this shouldn't be too much of an issue for deaf devices as the batch command of Delete Orphans will allow "half links" to be created by deleting responder records from deaf devices.  As a result, the command on a deaf device will simply clean up the half links.  (I hope that makes sense)