AD2: Fix Errors in Alpha Fault Processing; Add Better Notes; Fix Bug in Partition
Checking for faults using the Alphanumeric messages is really complicated.  The concept sounds simple, but building a test that doesn't get confused is difficult. I think this current system is more accurate, but a little slower to move from fault->ready.

I added a lot more comments to explain my logic since code alone is not really enough.

Also fixed a error in the partition child.