AD2: Better Restore of Cached Zone States on Reboot
Certain zone states, such as Bypass, can only be determined when they are set.  As a result, we need to cache and save the zone states through a reboot, otherwise we will change a zone from bypass->ready whenever we reboot.

This does leave open the possibility of incorrectly setting a zone state if the zone changed state while MisterHouse was down.