Add Support for Pushbullet.com Notifications
Pushbullet is similar to Pushover.  It allows you to send notification messages directly from MH to your devices.

A device can be an iOS or Android phone, a windows PC, or any Chrome browser.  All of the clients are free and the service is free.

Pushbullet has many features:
- Send notes, urls, files, lists, and addresses
- Add multiple devices and friends
- Choose which devices and friends to send a particular push to

Right now the code provides a user friendly mechanism to send notes.  A low level module also provides raw support for sending whatever you want to the Pushbullet API.  In future commits, I hope to make additional features easy to use.

The addition of this feature to MH and some of the code was inspired by George Clark's work on the Pushover support.